Move UI-related code from `QueryRunner` into `LocalQueries`
