+++ extensions/ql-vscode/src/extension.ts
@@ -115,10 +115,7 @@ import {
   QueryServerCommands,
   TestUICommands,
 } from "./common/commands";
-import {
-  getLocalQueryCommands,
-  showResultsForCompletedQuery,
-} from "./local-queries";
+import { LocalQueries } from "./local-queries";
 import { getAstCfgCommands } from "./ast-cfg-commands";
 import { getQueryEditorCommands } from "./query-editor";
 import { App } from "./common/app";
@@ -712,12 +709,6 @@ async function activateWithInstalledDistribution(
   void extLogger.log("Initializing query history manager.");
   const queryHistoryConfigurationListener = new QueryHistoryConfigListener();
   ctx.subscriptions.push(queryHistoryConfigurationListener);
-  const showResults = async (item: CompletedLocalQueryInfo) =>
-    showResultsForCompletedQuery(
-      localQueryResultsView,
-      item,
-      WebviewReveal.Forced,
-    );
   const queryStorageDir = join(ctx.globalStorageUri.fsPath, "queries");
   await ensureDir(queryStorageDir);
 
@@ -791,8 +782,10 @@ async function activateWithInstalledDistribution(
     ctx,
     queryHistoryConfigurationListener,
     labelProvider,
-    async (from: CompletedLocalQueryInfo, to: CompletedLocalQueryInfo) =>
-      showResultsForComparison(compareView, from, to),
+    async (
+      from: CompletedLocalQueryInfo,
+      to: CompletedLocalQueryInfo,
+    ): Promise<void> => showResultsForComparison(compareView, from, to),
   );
 
   ctx.subscriptions.push(qhm);
@@ -813,7 +806,8 @@ async function activateWithInstalledDistribution(
     cliServer,
     queryServerLogger,
     labelProvider,
-    showResults,
+    async (item: CompletedLocalQueryInfo) =>
+      localQueries.showResultsForCompletedQuery(item, WebviewReveal.Forced),
   );
   ctx.subscriptions.push(compareView);
 
@@ -848,6 +842,18 @@ async function activateWithInstalledDistribution(
     },
     true,
   );
+
+  const localQueries = new LocalQueries(
+    app,
+    qs,
+    qhm,
+    dbm,
+    cliServer,
+    databaseUI,
+    localQueryResultsView,
+    queryStorageDir,
+  );
+  ctx.subscriptions.push(localQueries);
 
   void extLogger.log("Initializing QLTest interface.");
   const testExplorerExtension = extensions.getExtension<TestHub>(
@@ -902,11 +908,7 @@ async function activateWithInstalledDistribution(
     ...databaseUI.getCommands(),
     ...dbModule.getCommands(),
     ...getAstCfgCommands({
-      queryRunner: qs,
-      queryHistoryManager: qhm,
-      databaseUI,
-      localQueryResultsView,
-      queryStorageDir,
+      localQueries,
       astViewer,
       astTemplateProvider,
       cfgTemplateProvider,
@@ -926,16 +928,7 @@ async function activateWithInstalledDistribution(
   }
 
   const queryServerCommands: QueryServerCommands = {
-    ...getLocalQueryCommands({
-      app,
-      queryRunner: qs,
-      queryHistoryManager: qhm,
-      databaseManager: dbm,
-      cliServer,
-      databaseUI,
-      localQueryResultsView,
-      queryStorageDir,
-    }),
+    ...localQueries.getCommands(),
   };
 
   for (const [commandName, command] of Object.entries(queryServerCommands)) {
